feat: tab usage and small style updates
